create notification component
